Revert "[FIX] 회원가입, 자유게시판 임시 커밋(미완성)"
